Move system models in their respective components dir.
